added browserify for adding npm packages
