+++ .gitignore
@@ -11,3 +11,4 @@
 /log/*
 !/log/.keep
 /tmp
+/node_modules
