+++ app/assets/javascripts/application.js
@@ -1,27 +1,12 @@
-// This is a manifest file that'll be compiled into application.js, which will include all the files
-// listed below.
-//
-// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
-// or any plugin's vendor/assets/javascripts directory can be referenced here using a relative path.
-//
-// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file.
-//
-// Read Sprockets README (https://github.com/rails/sprockets#sprockets-directives) for details
-// about supported directives.
-//
-//= require jquery
-//= require jquery_ujs
-//= require jquery
-//= require jquery-ui/widgets/datepicker
-//= require bootstrap-sprockets
-//= require turbolinks
-//= require react
+//= require_self
+//= require react-server
 //= require react_ujs
-//= require components
-//= require_tree .
 
-$( function() {
-	console.log('app.js')
-    $( "#datepicker" ).datepicker();
- } );
+// Setup React in global scope
+window.$ = window.jQuery = global.$ = require('jquery');
+var React = window.React = global.React = require('react');
+var ReactDOM = window.ReactDOM = global.ReactDOM = require('react-dom');
+var ReactDOMServer = window.ReactDOMServer = global.ReactDOMServer = require('react-dom/server');
+require( 'jquery-ujs' );
+require( 'fetch' );
+require( './components' );
