+++ app/assets/javascripts/components/Players.js
@@ -1,0 +1,11 @@
+// Your third party add-ons from react component
+// Define component
+import RinkSearch from './RinkSearch';
+
+export default class Player extends React.Component {
+  render(){
+    return (
+    	<RinkSearch />
+    );
+  }
+}
