--- app/assets/javascripts/components/Players.js.jsx
@@ -1,7 +1,0 @@
-var Players = React.createClass({
-	render: function() {
-		return(
-			<RinkSearch />
-		)
-	}
-})
