+++ app/views/layouts/application.html.haml
@@ -6,9 +6,6 @@
     = javascript_include_tag 'application', 'data-turbolinks-track' => true
     = csrf_meta_tags
     %meta{:content => "user-scalable=no, initial-scale=1.0, maximum-scale=1.0", :name => "viewport"}/
-    %link{:href => "//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css", :rel => "stylesheet"}/
-    %script{:src => "https://code.jquery.com/jquery-1.12.4.js"}
-    %script{:src => "https://code.jquery.com/ui/1.12.1/jquery-ui.js"}
 
   %body
     - if(user_signed_in?)
