+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "OpenHockey",
+  "dependencies": {
+    "axios": "^0.15.3",
+    "babel-plugin-syntax-async-functions": "^6.3.13",
+    "babel-plugin-transform-regenerator": "^6.3.18",
+    "babel-polyfill": "^6.3.14",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "babelify": "^7.2.0",
+    "browserify": "^12.0.1",
+    "browserify-incremental": "^3.0.1",
+    "es6-promise": "^3.0.2",
+    "fetch": "^0.3.6",
+    "jquery": "^2.1.4",
+    "jquery-ujs": "^1.0.4",
+    "moment": "^2.17.1",
+    "react": "^0.14.3",
+    "react-datepicker": "^0.41.1",
+    "react-dom": "^15.4.1"
+  },
+  "license": "MIT",
+  "engines": {
+    "node": ">= 0.10"
+  },
+  "devDependencies": {
+    "babelify": "^7.3.0",
+    "browserify-css": "^0.10.0"
+  }
+}
